detector/requirements.txt
pyyaml>=6.0
flask>=3.0
psutil>=5.9
requests>=2.31
matplotlib>=3.7
